+++ apps/shared/package.json
@@ -4,7 +4,8 @@
   "version": "0.0.0",
   "type": "module",
   "exports": {
-    ".": "./src/index.ts"
+    ".": "./src/index.ts",
+    "./billing": "./src/billing-types.ts"
   },
   "types": "./src/index.ts",
   "scripts": {
